move sharable items up a dir level
