+++ isaaclab_arena_environments/isaac_cap/__init__.py
@@ -2,3 +2,14 @@
 # All rights reserved.
 #
 # SPDX-License-Identifier: Apache-2.0
+
+"""Isaac Cap environments and shared support code."""
+
+__all__ = ["register_components"]
+
+
+def register_components() -> None:
+    """Register all available Isaac Cap components."""
+    from .registration import register_components as _register_components
+
+    _register_components()
