+++ isaaclab_arena_environments/isaac_cap/embodiments/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Embodiment families used by Isaac Cap environments."""
