+++ isaaclab_arena_environments/isaac_cap/embodiments/insertion_task/cameras.py
