+++ isaaclab_arena_environments/isaac_cap/embodiments/insertion_task/config.py
@@ -20,14 +20,16 @@ from isaaclab.managers import SceneEntityCfg
 from isaaclab.sim.schemas.schemas_cfg import ArticulationRootBaseCfg
 from isaaclab.utils.configclass import configclass
 
-from ..asset_factories import ROBOT_ON_CART_USD_PATH
-
 ARM_JOINT_NAMES = [f"fr3_joint{index}" for index in range(1, 8)]
 GRIPPER_JOINT_NAME = "left_driver_joint"
 END_EFFECTOR_BODY_NAME = "robotiq_base"
 # The authored driver upper limit is 51.5662 degrees.
 GRIPPER_CLOSED_ANGLE = math.radians(51.5662)
 _DROID_WORKING_HEIGHT_M = 1.35
+_ROBOT_ON_CART_USD_PATH = (
+    "omniverse://isaac-dev.ov.nvidia.com/Projects/nvblox/isaac_arena/newton_envs/cap_envs/gear_assembly/"
+    "assets/industrial__fr3_robotiq_2f85_on_cart/industrial__fr3_robotiq_2f85_on_cart.usda"
+)
 
 
 @configclass
@@ -37,7 +39,7 @@ class IndustrialFr3RobotiqSceneCfg:
     robot: ArticulationCfg = ArticulationCfg(
         prim_path="{ENV_REGEX_NS}/Robot",
         spawn=sim_utils.UsdFileCfg(
-            usd_path=str(ROBOT_ON_CART_USD_PATH),
+            usd_path=_ROBOT_ON_CART_USD_PATH,
             activate_contact_sensors=True,
             articulation_props=ArticulationRootBaseCfg(fix_root_link=False),
         ),
